Archive/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
Flask-Login
Flask-Mail
email-validator 
gunicorn
